Move Custom Object Mapper
