+++ vcell-core/pom.xml
@@ -682,11 +682,6 @@
 			<version>5.0.1</version>
 		</dependency>
 
-		<dependency>
-			<groupId>com.fasterxml.jackson.datatype</groupId>
-			<artifactId>jackson-datatype-jsr310</artifactId>
-			<version>${jackson-datatype.version}</version>
-		</dependency>
 
     </dependencies>
